+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version X.Y.Z (YYYYMMDDXX)
 
+- Implement store, retrieve, and delete functionality using the Moodle Filestore backend.
 - Add Moodle plugin CI for all supported Moodle versions